bvcs_backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
exceptiongroup==1.3.1
iniconfig==2.3.0
packaging==26.2
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.12.0
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
